# schemas/api/openapi.yaml
openapi: 3.1.0
info:
  title: ShakerProxy Control API
  version: 0.1.0-dev
  description: |
    Setup-safe API. Only operations implemented by this build are listed.

    Discover resources with GET /api/v1 and fetch this document from GET /api/v1/openapi.yaml.
    Every error, including unknown routes (404 not_found) and unsupported methods
    (405 method_not_allowed with an Allow header), uses the JSON shape
    {"error":{"code":"snake_case","message":"What went wrong and what to do next."}}.
    Request bodies are JSON (Content-Type application/json, optional charset=utf-8);
    decoding errors name the offending field.

    Authentication: sign in with POST /auth/login and send Authorization: Bearer <session_token>.
    Sessions slide (1 hour idle, 12 hours maximum); authenticated responses carry
    X-ShakerProxy-Session-Expires-At. Scoped API tokens (POST /auth/tokens) use the same header.

    Password confirmation: fields named "password" confirm the administrator. After any
    successful password check (including sign-in) the same session may omit "password" for
    10 minutes on routes marked "recent confirmation"; they answer 401 reauthentication_required
    when a password is needed. Routes marked "password every time" (network commit/confirm,
    deletion execution and retention, API token creation, enabling HTTPS decryption for every
    client, enabling decrypted HTTP body retention) answer 401 password_required without one.
servers:
  - url: /api/v1
paths:
...
  /system/ports:
    get:
      operationId: getServicePortPlan
      description: Read-only ownership plan for TCP/UDP 53, 443, 853, and management TCP/8443. A loopback-only systemd-resolved stub is preserved and is not treated as a wildcard conflict. This endpoint never changes a listener or enables DNS.
      security: [{bearerAuth: []}]
      responses:
        "200": {description: Fixed service-port ownership inventory, content: {application/json: {schema: {$ref: "#/components/schemas/ServicePortPlan"}}}}
        "401": {$ref: "#/components/responses/Error"}
        "503": {$ref: "#/components/responses/Error"}
  /system/connectivity-probe:
    post:
      operationId: probeConnectivity
      description: Session-only, explicitly initiated and rate-limited fixed-destination probes. HTTPS is dialed at 1.1.1.1 with verified TLS for one.one.one.one, so its result does not depend on DNS. TCP/53 tests only reachability and never sends a DNS query.
      security: [{bearerAuth: []}]
      requestBody: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false}}}}
      responses:
        "200": {description: DNS-independent connectivity comparison, content: {application/json: {schema: {$ref: "#/components/schemas/ConnectivityReport"}}}}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "429": {$ref: "#/components/responses/Error"}
        "503": {$ref: "#/components/responses/Error"}
...
  /event-query-snapshots:
    post:
      operationId: createEventQuerySnapshot
      description: Canonicalizes a traffic query, freezes relative time and friendly-name resolution, establishes an immutable ingestion watermark, and returns an actor-bound selection reference that expires in 1 to 60 minutes.
      security: [{bearerAuth: []}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/EventQuerySnapshotInput"}}}}
      responses:
        "201": {description: Frozen traffic population, content: {application/json: {schema: {$ref: "#/components/schemas/EventQuerySnapshot"}}}}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
        "422": {$ref: "#/components/responses/Error"}
        "503": {$ref: "#/components/responses/Error"}
  /event-query-snapshots/{snapshotID}:
    get:
      operationId: getEventQuerySnapshot
      description: Returns an unexpired query snapshot only to the actor that created it.
      security: [{bearerAuth: []}]
      parameters: [{$ref: "#/components/parameters/EventQuerySnapshotID"}]
      responses:
        "200": {description: Frozen traffic population, content: {application/json: {schema: {$ref: "#/components/schemas/EventQuerySnapshot"}}}}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
        "503": {$ref: "#/components/responses/Error"}
  /saved-views:
    get:
      operationId: listSavedViews
      description: Lists at most 100 personal views owned by the session actor plus visible shared views.
      security: [{bearerAuth: []}]
      parameters:
        - {in: query, name: scope, schema: {type: string, enum: [personal, shared]}}
        - {in: query, name: page, schema: {const: live-traffic}}
      responses:
        "200": {description: Bounded saved-view list, content: {application/json: {schema: {$ref: "#/components/schemas/SavedViewPage"}}}}
        "401": {$ref: "#/components/responses/Error"}
        "503": {$ref: "#/components/responses/Error"}
    post:
      operationId: createSavedView
      description: Canonicalizes and persists a new owned saved view at revision one.
      security: [{bearerAuth: []}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/SavedViewConfiguration"}}}}
      responses:
        "201": {description: Saved view created, content: {application/json: {schema: {$ref: "#/components/schemas/SavedView"}}}}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "503": {$ref: "#/components/responses/Error"}
...
  /saved-views/{viewID}:
    parameters:
      - {$ref: "#/components/parameters/SavedViewID"}
    get:
      operationId: getSavedView
      security: [{bearerAuth: []}]
      responses:
        "200": {description: Visible saved view, content: {application/json: {schema: {$ref: "#/components/schemas/SavedView"}}}}
        "404": {$ref: "#/components/responses/Error"}
    put:
      operationId: updateSavedView
      description: Owner-only optimistic revision update that appends an immutable version snapshot.
      security: [{bearerAuth: []}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/SavedViewUpdate"}}}}
      responses:
        "200": {description: Updated saved view, content: {application/json: {schema: {$ref: "#/components/schemas/SavedView"}}}}
        "403": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
    delete:
      operationId: deleteSavedView
      description: Owner-only optimistic deletion of a non-destructive view configuration.
      security: [{bearerAuth: []}]
      parameters:
        - {in: query, name: expected_revision, required: true, schema: {type: integer, minimum: 1}}
      responses:
        "204": {description: Saved view deleted}
        "403": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, description: Browser sessions authenticate every route. Expiring scoped API tokens authenticate only explicitly scoped read/automation routes and cannot manage credentials, forwarders, networking, deletion, exports, rules, or trust material.}
  responses:
    Error:
      description: Structured error response
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error:
                type: object
                required: [code, message]
                properties:
                  code: {type: string, pattern: "^[a-z0-9_]+$"}
                  message: {type: string, description: What went wrong and what to do next.}
  parameters:
...
    SavedViewID:
      in: path
      name: viewID
      required: true
      schema: {type: string, pattern: "^view-[a-f0-9]{32}$"}
    EventQuerySnapshotID:
      in: path
      name: snapshotID
      required: true
      schema: {type: string, pattern: "^qsnap-[a-f0-9]{32}$"}
...
  schemas:
...
    PortListener:
      type: object
      additionalProperties: false
      required: [transport, address, port]
      properties:
        transport: {type: string, enum: [tcp, udp]}
        address: {type: string, minLength: 1, maxLength: 64}
        port: {type: integer, minimum: 1, maximum: 65535}
        process: {type: string, minLength: 1, maxLength: 64}
    PortReservation:
      type: object
      additionalProperties: false
      required: [transport, port, purpose, intended_binding, state, action, listeners]
      properties:
        transport: {type: string, enum: [tcp, udp]}
        port: {type: integer, enum: [53, 443, 853, 8443]}
        purpose: {type: string, minLength: 1, maxLength: 128}
        intended_binding: {type: string, minLength: 1, maxLength: 64}
        state: {type: string, enum: [FREE, COEXIST, OWNED_EXPECTED, POTENTIAL_CONFLICT]}
        action: {type: string, enum: [AVAILABLE, PRESERVE_STUB_BIND_LAB_ADDRESS_ONLY, PRESERVE_SHAKERPROXY_OWNER, REVIEW_OR_REBIND_BEFORE_ENABLE]}
        listeners: {type: array, maxItems: 4096, items: {$ref: "#/components/schemas/PortListener"}}
    ServicePortPlan:
      type: object
      additionalProperties: false
      required: [schema, generated_at, systemd_resolved_stub, resolver_handling, reservations]
      properties:
        schema: {const: 1}
        generated_at: {type: string, format: date-time}
        systemd_resolved_stub: {type: boolean}
        resolver_handling: {type: string, enum: [NO_STUB_DETECTED, PRESERVE_STUB_BIND_LAB_ADDRESS_ONLY]}
        reservations: {type: array, minItems: 7, maxItems: 7, items: {$ref: "#/components/schemas/PortReservation"}}
    ConnectivityProbe:
      type: object
      additionalProperties: false
      required: [name, target, status, detail]
      properties:
        name: {type: string, enum: [dns-independent-https-ipv4, plain-dns-tcp-ipv4]}
        target: {type: string, enum: ["1.1.1.1:443", "1.1.1.1:53"]}
        status: {type: string, enum: [PASS, FAIL, UNKNOWN]}
        latency_millis: {type: integer, minimum: 0, maximum: 7000}
        detail: {type: string, maxLength: 160}
    ConnectivityReport:
      type: object
      additionalProperties: false
      required: [schema, generated_at, dns_independent_https, plain_dns_port_53, restricted_port_53, probes]
      properties:
        schema: {const: 1}
        generated_at: {type: string, format: date-time}
        dns_independent_https: {type: boolean}
        plain_dns_port_53: {type: boolean}
        restricted_port_53: {type: boolean, description: True only when fixed-IP HTTPS succeeds and fixed-IP TCP/53 fails.}
        probes: {type: array, minItems: 2, maxItems: 2, items: {$ref: "#/components/schemas/ConnectivityProbe"}}
...
    EventQuerySnapshotInput:
      type: object
      additionalProperties: false
      required: [query, sort]
      properties:
        query: {type: string, maxLength: 2048, description: Typed traffic filter. The server returns its canonical representation.}
        sort:
          type: array
          minItems: 2
          maxItems: 2
          prefixItems:
            - {type: object, additionalProperties: false, required: [field, direction], properties: {field: {const: occurred_at}, direction: {const: desc}}}
            - {type: object, additionalProperties: false, required: [field, direction], properties: {field: {const: record_id}, direction: {const: desc}}}
        expires_in_seconds: {type: integer, minimum: 60, maximum: 3600, default: 900}
    EventQuerySnapshot:
      type: object
      additionalProperties: false
      required: [schema, query_snapshot_id, canonical_query, sort, matched_count, count_relation, created_at, expires_at, dataset_watermark, snapshot_sha256, policy_version]
      properties:
        schema: {const: 1}
        query_snapshot_id: {type: string, pattern: "^qsnap-[a-f0-9]{32}$"}
        canonical_query: {type: string, maxLength: 2048}
        sort:
          type: array
          minItems: 2
          maxItems: 2
          prefixItems:
            - {type: object, additionalProperties: false, required: [field, direction], properties: {field: {const: occurred_at}, direction: {const: desc}}}
            - {type: object, additionalProperties: false, required: [field, direction], properties: {field: {const: record_id}, direction: {const: desc}}}
        query_anchor: {type: string, format: date-time, description: Present when a relative time predicate was frozen.}
        matched_count: {type: integer, minimum: 0, maximum: 100000}
        count_relation: {type: string, enum: [eq, gte], description: gte means matched_count is the bounded lower limit rather than an exact total.}
        created_at: {type: string, format: date-time}
        expires_at: {type: string, format: date-time}
        dataset_watermark: {$ref: "#/components/schemas/EventDatasetWatermark"}
        snapshot_sha256: {type: string, pattern: "^[a-f0-9]{64}$"}
        policy_version: {const: event-query-v1}
    EventDatasetWatermark:
      type: object
      additionalProperties: false
      required: [ingest_sequence, received_at, record_id]
      properties:
        ingest_sequence: {type: integer, minimum: 0, description: Authoritative monotonic database-assigned ingestion boundary.}
        received_at: {type: string, format: date-time}
        record_id: {type: string, pattern: "^[a-f0-9]{64}$"}
    SavedViewConfiguration:
      type: object
      additionalProperties: false
      required: [scope, name, page, time_behavior, sort, columns, pinned_columns, density, chart]
      properties:
        scope: {type: string, enum: [personal, shared]}
        name: {type: string, minLength: 1, maxLength: 96}
        description: {type: string, maxLength: 512}
        page: {const: live-traffic}
        canonical_query: {type: string, maxLength: 2048}
        time_behavior: {$ref: "#/components/schemas/SavedViewTimeBehavior"}
        sort: {type: array, minItems: 1, maxItems: 3, uniqueItems: true, items: {$ref: "#/components/schemas/SavedViewSort"}}
        columns: {type: array, minItems: 1, maxItems: 16, uniqueItems: true, items: {type: string, enum: [source, occurred_at, received_at, kind, device, network, protocol, service, bytes, confidence, capture, flow]}}
        pinned_columns: {type: array, maxItems: 3, uniqueItems: true, items: {type: string}}
        density: {type: string, enum: [comfortable, compact]}
        chart: {$ref: "#/components/schemas/SavedViewChart"}
    SavedViewTimeBehavior:
      type: object
      additionalProperties: false
      required: [mode]
      properties:
        mode: {type: string, enum: [query, rolling, absolute]}
        rolling_window: {type: string, enum: [15m, 1h, 6h, 24h, 7d]}
        start: {type: string, format: date-time}
        end: {type: string, format: date-time}
    SavedViewSort:
      type: object
      additionalProperties: false
      required: [field, direction]
      properties:
        field: {type: string, enum: [occurred_at, received_at, source, kind, device, source_ip, destination_ip, protocol, service, network_bytes, confidence]}
        direction: {type: string, enum: [asc, desc]}
    SavedViewChart:
      type: object
      additionalProperties: false
      required: [visible]
      properties:
        visible: {type: boolean}
        metric: {type: string, enum: [events]}
        interval: {type: string, enum: [1m, 5m, 15m, 1h]}
    SavedView:
      type: object
      additionalProperties: false
      required: [schema, id, scope, name, page, time_behavior, sort, columns, pinned_columns, density, chart, owner, last_editor, revision, created_at, updated_at]
      properties:
        schema: {const: 1}
        id: {type: string, pattern: "^view-[a-f0-9]{32}$"}
        scope: {type: string, enum: [personal, shared]}
        name: {type: string, minLength: 1, maxLength: 96}
        description: {type: string, maxLength: 512}
        page: {const: live-traffic}
        canonical_query: {type: string, maxLength: 2048}
        time_behavior: {$ref: "#/components/schemas/SavedViewTimeBehavior"}
        sort: {type: array, minItems: 1, maxItems: 3, uniqueItems: true, items: {$ref: "#/components/schemas/SavedViewSort"}}
        columns: {type: array, minItems: 1, maxItems: 16, uniqueItems: true, items: {type: string, enum: [source, occurred_at, received_at, kind, device, network, protocol, service, bytes, confidence, capture, flow]}}
        pinned_columns: {type: array, maxItems: 3, uniqueItems: true, items: {type: string}}
        density: {type: string, enum: [comfortable, compact]}
        chart: {$ref: "#/components/schemas/SavedViewChart"}
        owner: {type: string, minLength: 1, maxLength: 96}
        last_editor: {type: string, minLength: 1, maxLength: 96}
        revision: {type: integer, minimum: 1}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    SavedViewPage:
      type: object
      additionalProperties: false
      required: [schema, views]
      properties:
        schema: {const: 1}
        views: {type: array, maxItems: 100, items: {$ref: "#/components/schemas/SavedView"}}
    SavedViewUpdate:
      type: object
      additionalProperties: false
      required: [expected_revision, configuration]
      properties:
        expected_revision: {type: integer, minimum: 1}
        configuration: {$ref: "#/components/schemas/SavedViewConfiguration"}
...